+++ extensions/jpa/runtime/src/main/java/org/springframework/transaction/PlatformTransactionManager.java
@@ -1,0 +1,4 @@
+package org.springframework.transaction;
+
+public interface PlatformTransactionManager {
+}
