+++ extensions/jpa/runtime/src/main/java/org/springframework/transaction/support/TransactionTemplate.java
@@ -1,0 +1,9 @@
+package org.springframework.transaction.support;
+
+import org.springframework.transaction.PlatformTransactionManager;
+
+public class TransactionTemplate {
+    public PlatformTransactionManager getTransactionManager() {
+        throw new UnsupportedOperationException("getTransactionManager is not supported");
+    }
+}
